Disable doc generation in grpc gem compilation for FIPS builds
In https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/6172,
a FIPS build wil reinstall the gRPC gem with a patched version that
links against the system OpenSSL.

The `gem install` command installs 9 MB of documentation for the gRPC
gem. We don't need this documentation, so let's just disable it.

Changelog: changed