Fix worker processes not starting up due to 0 processes
The update to CINC 17.10 in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/6048
updated Ohai to v17.9.0. As a result of
https://github.com/chef/ohai/pull/1454, `lscpu` is used to determine
the number of real cores, total CPUs, and threads per core. However,
on some older platforms, threads per core is 0, resulting in the total
number of virtual CPUs being 0. This would cause NGINX to spin up with
0 cores.

To prevent startup failures, we now attempt to fallback to the number
of cores. Failing that, for NGINX we set the minimum worker processes
to 1.

We also convert to using `dig` in case the CPU data is not available
for some reason.

Relates to:

* https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6895
* https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6896

Changelog: fixed