Reload Consul only if the service is enabled
If for some reason an EE installation with Consul is downgraded to a
CE installation, `gitlab-ctl reconfigure` will fail with an obscure
message:

```
NoMethodError occurred in delayed notification: undefined method
`run_action' for "execute[reload consul]":String
```

In a CE build, the Consul recipes are not loaded so `reload consul` is
not available.

To avoid this error, we can use the `Services.enabled?` helper, which
will verify that Consul is in the list of available services in the
first place.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6934

Changelog: fixed