Limit concurrency of package publish jobs using resource_group
The package publish jobs can easily overwhelm the package server. When
all the publish jobs are executed at the same time, most of them fail
due to 5xx responses from the package server.
This commit uses a resource_group to limit the number of concurrent
package publish jobs that can be executed. They will be executed one
at a time per commit. Note that this will not restrict concurrency
for pipelines running on different commits.