Remove svlogd filter from consul configuration
Consul changed the log line format in version 1.7.0 - and at the same
time added in thresholds that made the logging less noisy. This
removes the svlogd filter that we added in !1946 as it is no longer
matching the log line, and is no longer necessary to suppress repeated
log entries.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6589

Changelog: changed