Move KAS mirror on DEV to the proper namespace
In order to properly build on CNG with all the optimizations, projects
should have the same namespace on canonical and dev, with the only
distinction of gitlab-org that should be gitlab on dev.

This commit will reflect the above change also on omnibus.

Related to https://gitlab.com/gitlab-org/cluster-integration/gitlab-agent/-/issues/300