Praefect: Deprecate DB metrics configuration
The `prometheus_exclude_database_from_default_metrics` configuration is
used to determine whether Praefect DB metrics are exported on the
default `/metrics` endpoint or on a separate `/db_metrics` endpoint.
Since scraping DB metrics requires expensive queries this configuration
is defaulted to true enabling separate configuration of scraping
intervals. In GitLab v15.5 this configuration is deprecated and in v16.0
removed. This change updates the deprecations list for this
configuration.

Changelog: deprecated