Allow output of Rails migrations in reconfigure step
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/5971
disabled output of steps in stderr/stdout to avoid leaking
passwords. This change also applied to suppressing the output of Rails
migrations, which can fail due to a variety of reasons. This made it
harder to debug upgrade failures.

`gitlab-rake db:migrate` should not leak any sensitive information, as
it is usually just prints out the status of the migrations and error
messages if one fails. This commit re-enables the output to ease
debugging.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6801

Changelog: changed