Support for 'track-repositories' Praefect command
Add the `track-repositories` command to `gitlab-ctl praefect` for easier
access. This allows many repositories to be tracked in a single batch by
reading the repositories to track from a file.

This also improves the formatting of the help text for `gitlab-ctl
prafect`. All argument descriptions now start on the same line as their
flags and are aligned. To re-use the same help text between
`track-repository` and `track-repositories`, we use the string as
a template that we add an arbitrary level of indentation to.

Links to documentation have also been fixed to reflect changes in
page anchors during recent updates.

Changelog: added