Update gpgme dependencies to match ruby-gpgme dependencies
https://github.com/ueno/ruby-gpgme/issues/154 upgraded ruby-gpgme to:

* libgpg-error 1.46
* libassuan 2.5.5
* gpgme to 1.18.0

ruby-gpgme was updated in GitLab Rails:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105155

libgpg-error v1.4.6 no longer installs `gpg-error-config` by default.
`gpg-error-config` has been deprecated in favor of `gpgrt-config`
(https://dev.gnupg.org/T5683), but libassuan and gnupg currently still
use `gpg-error-config` if it is present. Older systems, such as Amazon
Linux 2, have a version that is too old, so we need to ensure a recent
version of `gpg-error-config` is installed via
`--enable-install-gpg-error-config`.

To ensure that the compiled version of `gpg-error-config` is used,
specify `--with-libgpg-error-prefix` in gnupg and libassuan packages.

Changelog: other