Use PIPELINE_TYPE logic for trigger jobs
We could not originally use `PIPELINE_TYPE` for triggered jobs because
`inherit:variables` was set to false. If enabled, then passing the
`PIPELINE_TYPE` variable to the downstream pipeline would interfere with
the workflow rules calculation downstream.

Here, the passed variables are specified with each `if` clause to avoid
interference with the workflow rules calculation.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7225

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7315

Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>