git: Make the Gitaly Git version override version-neutral
With 89950647c (Set GIT_REPO_URL, GIT_VERSION_2_37_1, GIT_VERSION_2_38,
2022-11-11), we have introduced the logic to allow overriding the
bundled Git versions that Gitaly is using. Back then we had been hard
coding the specific versions that could be overridden. But as these
bundled Git versions are under the control of Gitaly and change over
time those are bound to become out-of-date.

In fact, they already are out-of-date as Gitaly only distributes the Git
v2.38 release train. Furthermore, Gitaly will soon add a second release
train with Git v2.39, which would require another new variable override.

Refactor the logic to be version-neutral so that we just forward all
environment variables that have the `GITALY_GIT_VERSION_` prefix to the
build environment.