Move QA image release to a different stage
This is done temporarily to test releasing of QA images to Dockerhub in
the GitLab rails pipelines. By moving it to a different stage, we
prevent release-tools from automatically playing it, while retaining the
ability to play it ourselves if the job in GitLab rails pipeline fails
for some reason.

Related MR in GitLab rails project - https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105177

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7333

Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>