Disable zstd decompression with libmagic
ruby-magic updated to libmagic v5.44, which added support for zstd
decompression support. Unlike most of our other builders, it appears
that Amazon Linux needs `libzstd` to support `dnf`.

With ruby-magic v0.6.0 in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/114641, we can
use the `--with-magic-flags` to disable this functionality outright
since this system dependency may not be safe to use.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7672

Changelog: changed