Fix SSH host key generation on CentOS 7
The version of ssh-keygen shipped with CentOS 7 does not support
the ability to save SSH host keys to an alternate path using the `-f`
option with the `-A` option. Keys will always save to `/etc/ssh`.

To fix this, manually call ssh-keygen for each key that we want to
generate. On FIPS systems, it's not guaranteed that ed25519 or other
keys can be generated, so we skip over any failures in `ssh-keygen`
and only check at the end to see if any keys were successfully
generated.

Note that the host key generation is only performed if no host keys
were detected in the directory in the first place.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7671

Changelog: fixed