Disable use of the system ffi gem
Prior to
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/6777,
the ffi development headers were not present in the ruby_docker image.
The ffi gem will use the system ffi if `ffi.h` is in the include
path. This causes issues as different operating systems include
different versions of libffi. Disabling the use of the system ffi gem
will cause the ffi gem to statically build its own version, which
what was done before.