Use CI image name in gems cache key
https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7736 revealed
that we use Debian `bullseye` to build all native gems for the
specs. Since native gems may depend on system libraries or specific
versions of glibc, we should adjust the gem caching for the specs to
depend on the builder image.

This commit modifies the `spec_template` to use a cache key dependent
on the OS. The cache policy is also the default pull-push to avoid
needing a separate CI job to update this cache.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7737