Drop bundler software definition
In https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/1265
we pinned the version of bundler to ensure the version would not
quietly change. However, this way of pinning bundler is no longer
needed with Bundler v2.3+, which uses the bundler version specified in
`Gemfile.lock`: https://bundler.io/blog/2022/01/23/bundler-v2-3.html.

With
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/6843, we
upgraded the version of RubyGems to v3.4.12, which ships with bundler
v2.4.12. This means that all `bundle install` invocations will use the
bundler version dictated by `Gemfile.lock`, so the bundler software
definition is redundant and unnecessary.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7775

Changelog: other