Add workaround for gprc native extension not compiling
On platforms that need to compile the grpc native extension, there is
a bug introduced in grpc v1.48.0 that causes the native extension to
include re2 headers from `/opt/gitlab/include` instead of
`vendor/re2`. To work around that problem, add `vendor/re2` in the
include path.

Relates to:

* https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7815
* Relates to https://github.com/grpc/grpc/pull/32580

Changelog: fixed