Add support for enabling requirepass in Sentinel configuration
In https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/6921,
we're adding support for configuring a password to authenticate with
Redis Sentinel. This commit adds support for the `requirepass` config
option so that the Omnibus-provided Sentinel can require a password.

If `sentinel['password']` is set, the Sentinel instance passwords
default to this password to make it easy to enable this option.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/235938

Changelog: added