Build PgBouncer from Git source
This makes it possible to automate the dependencies updates since we
don't have to update a checksum. Previously the release source tarball
was downloaded.

This also adds some redundancy if the GitHub binary can't be
downloaded since we now have mirrors for this repository.

This came out of
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/6933.

Changelog: changed