Fix reconfigure failing when Sentinel TLS is only enabled
If the unencrypted Sentinel port is disabled and only TLS enabled,
`gitlab-ctl reconfigure` previously failed because it attempted to
talk to port 0, which is an invalid port.

This commit fixes this problem by switching to TLS if is the only
option. In addition, server and client certificates need to be
provided to `redis-cli` if required.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6627

Changelog: fixed