Use XZ compression for RPM packages
This is a follow-on to
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/7128
where we enabled XZ compression for DEB packages. Unfortunately, Omnibus
didn't do threaded compression for RPM packages because of the way it
was setting `_binary_payload`. That was causing XZ compression in
particular to take 2-3 times longer than GZip, so we could not merge
that change at the same time. With
https://gitlab.com/gitlab-org/omnibus/-/merge_requests/48 merged into
our fork, Omnibus will use threaded compression for RPM by default and
we can enable XZ compression for RPM packages as well.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8222

Changelog: changed