Add git-filter-repo
Gitaly is in the process of adding the ability to rewrite repository
history. To do this, we will use `git-filter-repo`, built by a
prominent contributer to the Git project. The script becomes available
as a Git subcommand when it is placed in $PATH.

Install the script to `/opt/gitlab/embedded/bin` via `pip3` and add
patches for all three of its license files.

Changelog: added