Create symlinks in /etc/ssh pointing to /etc/gitlab for SSH private keys
The current gitlab Docker image creates SSH host keys during build in
/etc/ssh when installing openssh-server. These keys are not used.
Instead, the SSH host keys for Gitlab are generated by the wrapper
script and stored in /etc/gitlab.

Previously, the wrapper script overwrote the unused SSH public host keys
in /etc/ssh with symlinks pointing to the keys used in /etc/gitlab. The
SSH private keys were left untouched. Thus, there were SSH private keys
in /etc/ssh that were not used and the corresponding public keys in
/etc/ssh were symlinks to /etc/gitlab and the private and public keys
did not match.

Since this is could be confusing for adminstrators, this commit alters
the wrapper script so that the unused SSH private keys in /etc/ssh are
replaced with symlinks to the private keys used in /etc/gitlab.