Correctly specify where CA store comes from
CAcert.org does not make a collection of trusted root CAs.
GitLab Omnibus uses Curl's version of Mozilla's trust store:
https://gitlab.com/gitlab-org/omnibus-gitlab/-/blob/e20cbf834fbc3d0ec6e07c9f6c107f5df8a3eb87/config/software/cacerts.rb#L35

I think this error resulted from confusing the `cacert.pem` file (which is the name for any file that contains a collection of CAs), and the `CAcert.org` non-profit (which is a non-profit CA).

Signed-off-by: syvb <me@iter.ca>