Add active directory + smart card settings to gitlab.yml
An upcoming change in GitLab will introduce two new settings in the
`ldap:` -> `servers:` subsections, to allow authenticating users via
smart card against LDAP servers that are Active Directory based. The
configuration is per-server-specific, so should be placed in the same
configuration block as other LDAP settings.

This change exposes the new settings and adds defaults to omnibus. The
`smartcard_ad_cert_format` setting should default to `nil` / off, and
administrators must opt-in to the new behavior by configuring a value
for it.

Changelog: added