Accept multiple bind addresses in Redis config
As specified in https://redis.io/docs/management/config-file/, Redis
can bind to multiple addresses with a space-separated field. Previously
attempting to do this without setting `gitlab_rails['redis_host']`
would fail because a URI could not be built with a space in the
hostname.

This commit now splits the string with the space and picks the first
address as the default Redis host.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8465

Changelog: added