Make Redis info methods used in libraries class methods
Certain methods in the Redis Helper are to be used in libraries also,
where the node object is unavailable. In those scenarios, the methods
have to work with the Gitlab object.

Make those methods class methods that take a config hash as argument.
Also create instance methods with same name, which call these class
methods and pass the node object as argument.

Signed-off-by: Balasankar 'Balu' C <balasankar@gitlab.com>