Fix patroni no longer working with update to ydiff 1.3
ydiff 1.3.0 was released on April 2, 2024
(https://pypi.org/project/ydiff/#history), and the patroni install
quietly updated to ydiff 1.3.0 from 1.2.0.

https://github.com/zalando/patroni/blob/v3.0.1/requirements.txt#L15
allows any version of ydiff above 1.2.0, but as described in
https://github.com/zalando/patroni/blob/634b44ee0586f063033074806a42b534a354cbff/docs/releases.rst#L16
ydiff 1.3 requires the cdiff module.

To fix this problem, pin the version of ydiff to 1.2.0 until
we can upgrade patroni.

Changelog: fixed