redis: Fix password auth with UNIX domain sockets
Previously if a Redis instance listened on a UNIX socket but a
password were set, GitLab Rails would not be able to authenticate.
This occurred because the UNIX URL doesn't contain a password.

Both Ruby and Go Redis clients support URLs in the form:

unix://<user>:<password>@</path/to/redis.sock>?db=<db_number>

Relates to work started in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/2194

Changelog: fixed