Pin parser dependency in chef-bin
Installing chef-bin via `gem install` makes all of its dependencies floating,
including the `parser` gem. Because the `parser` gem is *also* used by other
Ruby modules, the floating dependencies can cause failures, such as a recent
problem that caused `gitlab-ctl` to fail because `inspec-core` depends on
`parser` as well. This is a temporary fix to correct the floating dependency
on `parser` until a more permanent fix is in place.