Upgrade some Omnibus-related gems
In https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/7362,
the `Gemfile.lock` was locked to a number of gem versions present
around January 2024.

As seen in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/7584#note_1897220176,
we have been pulling in more recent versions of these gems for
recent patch releases, so let's use these updated versions
so that we minimize the diff.