Drop Cinc EOL warning
Previously users would see this warning at the end of a reconfigure:

```
WARN: This release of Cinc Client became end of life (EOL) on May 1st 2024
```

This warning started to show up on a May 1, 2024, but the admin has no
control over what version of Cinc is used. Disable this warning since
it doesn't provide any useful information.

Changelog: changed