redis: Fix handling of passwords with a space
Previously we used `CGI.escape(password)`, but this escapes spaces as
`+` instead of `%20`. However, RFC 3986 says the password should be
percent-encoded. In fact, while the Ruby Redis client appears to
connect fine with the `+`, the Go client doesn't work with that.
Update the helper to encode passwords according to RFC 3986.

Changelog: fixed