Ensure Omnibus GitLab gems are recompiled for older platforms
The ffi gem just released a precompiled gem for `aarch64-linux-musl`,
which is built with Ubuntu 20.04 with glibc 2.27. As a result, older
systems, such as Amazon Linux 2, need to enable `force_ruby_platform
true` to ensure any native gems continue to work. This will
be needed if the `ffi` gem is updated to 1.17.0.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8570