Fix stable version tag identification for 16.1.x releases
In https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/7681
we found that the `upgrade-minor` QA job was attempting to upgrade
from 16.10 to 16.11 for a merge request targeting `16-1-stable`.  This
occurred because the version matching check only found a stable branch
that started with `16.1`, so `16.11.3` was a valid match.

Fix this by tightening the match to include a period, so that `16.1.`
will match `16.1.7`.

Changelog: fixed