Raise default PostgreSQL shared buffers minimum to 256 MB
Previously the Docker container set the PostgreSQL shared buffers to a
minimum of 1 MB to avoid using the default of 25% of the node's
memory. While that meager value worked previously, this value is not
only overly conservative but it is now insufficient for background
migrations that have a large number of partitioned tables.

From
https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8682#note_2058612975
we see that the 2 years of `audit_events` needs about 20 MB just to
query the table metadata. The 1 MB value failed with `no unpinned
buffers available` because PostgreSQL needs more shared buffers for
query planning.

Raise the default to 256 MB to provide enough headroom for future
tables and to ensure PostgreSQL effectively caches data. Update the
documentation in the template accordingly.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8682

Changelog: fixed