Downgrade OpenSSL version to 1.1.1
This reverts the changes in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/7797
because this breaks LDAP and other client integrations that don't
support TLS 1.2:
https://github.com/openssl/openssl/issues/13299#issuecomment-722124937

TLS 1.0 and 1.1 are supported by OpenSSL 3 if
`SSLContext#security_level=` is set to 0, but as far as I can tell
there's no simple way to ensure that this is enabled by default right
now. So until we do this properly or have a proper deprecation process
for TLS 1.1 and 1.0, we probably need to stay on OpenSSL 1.1.1.