Install OpenSSL gem after Ruby
For FIPS packages for OSs that ship OpenSSL 3+ to work properly, the
changes made as part of https://github.com/ruby/openssl/issues/603 is
required. This is available in v3.2.0 of the OpenSSL gem. Installing it
after installing Ruby will cause Ruby to use it instead of the default
OpenSSL gem version.

Signed-off-by: Balasankar 'Balu' C <balasankar@gitlab.com>