Recompile native gems for SLES 15.2
As described in
https://github.com/rake-compiler/rake-compiler-dock/issues/122, x86
precompiled native gems for a long time have used an image based on
CentOS 7, which uses glibc 2.17.

With https://github.com/rake-compiler/rake-compiler-dock/pull/126, the
image has been updated to Ubuntu 20.04, which uses glibc 2.29.

SLES 15.2 uses glibc 2.26, so we need to recompile Ruby native gems
for them to work. Failure to do so results in a `LoadError`:

```
LoadError: /lib64/libc.so.6: version `GLIBC_2.28' not found
```

See
https://gitlab.com/gitlab-org/build/omnibus-gitlab-mirror/-/jobs/8861917454
as an example.

This is a follow-up to
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8087.

Changelog: changed