Revert "Merge branch 'update-pg-feb-2025' into 'master'"
This reverts merge request !8201

The merge request had a typo in the title and body which
create a broken changelog. After consultation with Skarbek
and Mayra, we decided to do a revert and re-apply to
fix things and delete the bad record from the generated
changelog on March 19th/20th.