Fix libdir configured in OpenSSL 3
When building Gitaly's Git version with `USE_MESON` we use Meson instead
of its old Makefile. Doing so causes a linker error though:

    cc  -o git-remote-http git-remote-http.p/http.c.o git-remote-http.p/http-walker.c.o git-remote-http.p/remote-curl.c.o -L/opt/gitlab/embedded/lib -I/opt/gitlab/embedded/include -Wl,--as-needed -Wl,--no-undefined -Wl,-rpath,/opt/gitlab/embedded/lib -O3 -fstack-protector -O3 -D_FORTIFY_SOURCE=2 -fstack-protector -Wl,--start-group libgit.a libgit-version.a libcommon-main.a /opt/gitlab/embedded/lib/libz.so -pthread -liconv /opt/gitlab/embedded/lib/libpcre2-8.so /opt/gitlab/embedded/lib/libcurl.so -lresolv /usr/lib/gcc/x86_64-linux-gnu/11/../../../x86_64-linux-gnu/libssl.so /usr/lib/gcc/x86_64-linux-gnu/11/../../../x86_64-linux-gnu/libcrypto.so -liconv -lresolv -liconv -lresolv -liconv -lresolv -Wl,--end-group
    /usr/bin/ld: /opt/gitlab/embedded/lib/libcurl.so: undefined reference to `SSL_get0_group_name@OPENSSL_3.2.0'
    collect2: error: ld returned 1 exit status

Note how the Meson has decided to pick up libcrypto and libssl from the
system path even though libcurl itself was linked against OpenSSL in our
own embedded path. This linker error only occurs when using OpenSSL 3 by
setting the `OPENSSL_VERSION=3.4.1` variable -- with OpenSSL 1 things
work just fine.

This error is somewhat puzzling: we do set up the `PKG_CONFIG_PATH`, so
we should be able to find the expected OpenSSL libraries. And the fact
that we manage to pick up libcurl indicates that this works as expected.
But when executing the pkg-config(1) command manually one does get a
lead:

```
$ PKG_CONFIG_PATH=/opt/gitlab/embedded/lib/pkgconfig/ pkg-config --libs openssl
-L/opt/gitlab/embedded -lssl -lcrypto
```

While this looks correct at first, note how we're missing the "/lib"
suffix in the linker search path. And indeed, when one inspects the
installed pkg-config file of OpenSSL you can see the following line:

```
prefix=/opt/gitlab/embedded
exec_prefix=${prefix}
libdir=${exec_prefix}
```

The `libdir` is indeed missing the "/lib" directory.

As it turns out, the root cause is a change in behaviour in OpenSSL
3.3.1 itself: while it previously worked alright to pass an absolute
path to `--libdir`, it now isn't anymore [1]. Instead, one is expected
to pass a relative path, and when one doesn't then the result is subtle
breakage as we have observed here.

Fix the issue by instead specifying the libdir relative to the prefix.

[1]: https://github.com/openssl/openssl/issues/25249