Fix compilation of Git with Meson in FIPS environments
When compiling Git with Meson we have to set up the "PKG_CONFIG_PATH" so
that we're able to locate our libraries. In a FIPS-enabled environment
though we have to adapt the "PKG_CONFIG_PATH" so that the FIPS-enabled
OpenSSL library is preferred over our self-compiled version. This is
done by prepending `#{OpenSSLHelper.pkg_config_dirs}` to the environment
variable.

This doesn't work as expected though as the helper will typically return
a pkg-config directory in the "/usr" prefix. The end result is that we
now prefer _all_ libraries from that prefix instead of using our own
versions installed into "/opt/gitlab/embedded".

Fix the issue by creating a separate pkg-config directory and copying
the OpenSSL pkg-config files into it. Like this we can prepend the
override directory to our "PKG_CONFIG_PATH" and only override those
specific libraries.