Use zlib-ng instead of zlib
zlib-ng is a replacement for zlib with optimizations for "next
generation" systems. As such it provides significant performance
improvements in certain use cases. One of these usecases is Git, which
uses zlib compression for its objects. Benchmarks have shown that
switching over to zlib-ng gives us performance improvements for around
25-30%.

Git v2.48 has learned to use zlib-ng as an alternative backend, so we
could switch over to it for Git explicitly. But we can go even further:
zlib-ng provides a compatibility layer that makes it fully compatible
with the zlib interface. So instead of only switching over Git, we can
enable this compatibility layer and use it for all of our components.
This doesn't only have a performance benefit, but it also allows us to
not carry both zlib and zlib-ng.

Add zlib-ng at v2.2.4 and enable the compatibility layer. We will switch
over to this new dependency in the next commits.