Restore frame pointers in Git
With the migration to Meson to compile Git we have lost frame pointers.
This is because the CFLAGS that we write into Gitaly's "config.mak" file
get appended via `GIT_APPEND_BUILD_OPTIONS`, which is only effective for
Makefile-based builds.

Fix this by adding `-fno-omit-frame-pointer` to Meson via the `CFLAGS`
environment variable.

Changelog: fixed