config/git: Fix misdetected shell path
We have received error reports from users that they cannot mirror
repositories anymore. The error is always a variant of the following:

    fatal: cannot exec 'ssh -oIdentitiesOnly=yes -oIdentityFile=/tmp/gitaly-ssh-invocation2473534902/ssh-key -oStrictHostKeyChecking=yes -oCheckHostIP=no -oUserKnownHostsFile=/tmp/gitaly-ssh-invocation2473534902/known-hosts': No such file or directory
    fatal: unable to fork

Important is the "No such file or directory" error here, which indicates
that we have hit `ENOENT` on fork. Furthermore, we see errors like the
following, as well:

    remote: fatal: cannot exec 'exit 0 #': No such file or directory

Again, we fail because of `ENOENT`.

At first it's not quite clear where this error may come from. But what's
common between those two is that Git tries to execute commands with the
configured shell. Seemingly, we're unable to locate the shell altogether
and thus fail to execute the command.

The shell used by Git is set up at build time. In the Makefile it's
always set to "/bin/sh", whereas Meson tries to autodetect the location
of the shell via a "PATH". We have recently switched to Meson, and
indeed the shell path as shown by `git vars -l` isn't "/bin/sh" anymore
but instead "/usr/bin/sh".

It is unclear at this point in time why this should cause any problems.
Tests with both Amazon Linux and Ubuntu 20.04 images show that the new,
auto-detected shell path exists as expected. But it is possible that
customers have environments where this path does _not_ exist, and that
would lead to the observed error.

Override the shell path to be "/bin/sh". This location is expected to
always exist as specified by the POSIX standard, so it should be a safe
assumption to make.

Changelog: fixed