Fix NGINX modules not always building with the right tag
If the Omnibus build cache were present, previously the
`ngx_security_headers` and `nginx-module-vts` components would often
be built with `master` instead of tagged versions. This happened
because Omnibus' `GitFetcher` clones the repositories and only checks
out the specific version when building the component.  The `nginx`
component assumes that the Omnibus source cache has the right version,
but if nothing changed with `ngx_security_headers` and
`nginx-module-vts` then the `git checkout` will not occur.

To avoid this trouble with `GitFetcher`, we need to do several things:

1. Copy the extracted source into a subdirectory into
`/opt/gitlab/src/nginx_modules`.

2. Point NGINX to these installed directories.

3. Exclude the `/opt/gitlab/src` from the final package.

The Omnibus build cache will cache the contents of `/opt/gitlab/src`
in Git, and so the right versions should always be used.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/9066

Changelog: fixed