Fix deprecation check failing on nil values
This is a regression in GitLab 18.2 from due to
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8470.
Previously a blank default version were provided to `check_config`.  Now
the value can be `nil` if the JSON nodes file does not have the value.

Fix this by handling this `nil` value gracefully.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/9217

Changelog: fixed